massively cleaned up
